2d array with master entity heap now compiles and runs
Added a heapsort mechanism, to act on a master entity list. This acts inside the Battle class and acts on an ArrayList of Entities currently on the field. (See comments in the constructor for details.) Also added some public getter methods, especially for Character & Entity objects. The code at this point should compile and run without error. Let me know if there are any issues. (Future developments might be to better encapsulate the data / methods. I also need to test the move() capabilities more, as I have not tested them yet, and I am also interested in adding different sub-classes to represent different types of units.)