Graphics Tests (committing some fun demos)
GraphicsTests is a program where you can copy a cell color to another cell. GraphicsLines is a demo in which you can click to generate a new random coloring of the grid. Try resizing the GraphicsLines example to have the window automatically generate random colorings at a remarkably fast speed.