Revised distance formula; Battle now compiles.
Have revised the distance formula so that the calculation in moveTest() is exact. (There were edge cases in the previous formula in which the hypotenuse was sufficiently small that the program thought a square was in movement range, when in reality is shouldn't have been. This error has been fixed by making distance an int, since all squares are discrete. Also, Battle.java now compiles as needed.