Added Timer; Improved movement functionality.
Added a Timer capability; an event (e.g. animation, which I expect to work on after getting menus working) can occur at a specific timestep on its own, without explicit user input. Also the movement functionality is more accurate, as a unit now cannot move onto an ocean square or to a tile that is out of bounds. (A precondition for this is that characters do not SPAWN on an ocean tile. Should this occur in the current form of the game, that unit will not be able to move. (We can fix this easily by making some units fly, or by simply never spawning units there in the first place.)