Reorganized TurnBasedSystem in order to add a spawning mechanism
Ensured that units do not spawn on ocean tiles, regardless of the map choice. This was done in a rather rudimentary manner, by forcing enemy units to spawn on the other side of the map as player units (a more nuanced system is in progress, but now at least it is impossible for units to spawn on ocean tiles, which would render that immotile). Also added methods to set an Entity position (in order to make this spawning behavior work). Updated README.