Introduce an alternative halt menu GUI
This new alternative arrangement provides a consistent space,
on the right-hand side of the screen, for any menu,
like the halt menu for when a Unit finishes moving.
To do this, an arrayList of "HaltMenuOption" objects
is stored as a "menu." This design allows the menu to have variable length,
thus only containing the valid, legal options to the user
at a given time.