Refine the side-menu options
This commit adds rudimentary options for drop/pass/etc. at the end of its turn.
It also modifies the Character class to incorporate a basic list of items (and thus adds Item.java as a Java interface).
More work is needed in the UI for allowing the user to choose a target.